[OF-1672] feat: Moved Web code to Common
