+++ Library/include/CSP/Systems/Quota/Quota.h
@@ -20,6 +20,7 @@
 #include "CSP/Common/Map.h"
 #include "CSP/Common/SharedEnums.h"
 #include "CSP/Common/String.h"
+#include "CSP/Common/Web/HTTPResponseCodes.h"
 #include "CSP/Systems/WebService.h"
 
 namespace csp::services
