+++ Library/include/CSP/Systems/Settings/SettingsCollection.h
@@ -22,6 +22,7 @@
 #include "CSP/Common/SharedEnums.h"
 #include "CSP/Common/String.h"
 #include "CSP/Common/Variant.h"
+#include "CSP/Common/Web/HTTPResponseCodes.h"
 #include "CSP/Systems/WebService.h"
 
 #include <functional>
