+++ Library/include/CSP/Systems/Spaces/Space.h
@@ -20,6 +20,7 @@
 #include "CSP/Common/Map.h"
 #include "CSP/Common/SharedEnums.h"
 #include "CSP/Common/String.h"
+#include "CSP/Common/Web/HTTPResponseCodes.h"
 #include "CSP/Systems/Spatial/SpatialDataTypes.h"
 #include "CSP/Systems/SystemsResult.h"
 #include "CSP/Systems/WebService.h"
