+++ Library/src/Common/Web/HttpAuth.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "Web/HttpAuth.h"
+#include "Common/Web/HttpAuth.h"
 
 #include <assert.h>
