+++ Library/src/Common/Web/HttpAuth.h
