+++ Library/src/Common/Web/HttpProgress.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "Web/HttpProgress.h"
+#include "Common/Web/HttpProgress.h"
 
 namespace csp::web
 {
