+++ Library/src/Common/Web/HttpResponse.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "Web/HttpResponse.h"
+#include "Common/Web/HttpResponse.h"
 
 namespace csp::web
 {
