+++ Library/src/Common/Web/Json.cpp
@@ -13,28 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#pragma once
-
-#include "CSP/Common/CancellationToken.h"
-#include "Common/DateTime.h"
-#include "Services/PrototypeService/AssetFileDto.h"
-#include "Web/WebClient.h"
+#include "Common/Web/Json.h"
 
 namespace csp::web
 {
 
-class RemoteFileManager
-{
-public:
-    RemoteFileManager(csp::web::WebClient* InWebClient);
-    ~RemoteFileManager();
-
-    void GetFile(
-        const csp::common::String& FileUrl, csp::services::ResponseHandlerPtr ResponseHandler, csp::common::CancellationToken& CancellationToken);
-    void GetResponseHeaders(const csp::common::String& Url, csp::services::ResponseHandlerPtr ResponseHandler);
-
-private:
-    csp::web::WebClient* WebClient;
-};
+// Todo :- More complex types will be handled here with template specialisations
 
 } // namespace csp::web
