+++ Library/src/Common/Web/POCOWebClient/POCOWebClient.h
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include "Web/WebClient.h"
+#include "Common/Web/WebClient.h"
 
 #include <Poco/Net/HTTPCookie.h>
 #include <Poco/Net/HTTPSClientSession.h>
