+++ Library/src/Common/Web/Uri.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "Web/Uri.h"
+#include "Common/Web/Uri.h"
 
 #include "Common/NumberFormatter.h"
 #include "Common/StlString.h"
