+++ Library/src/Multiplayer/Conversation/Conversation.cpp
@@ -16,8 +16,8 @@
 #include "CSP/Multiplayer/Conversation/Conversation.h"
 #include "CSP/Systems/Assets/Asset.h"
 
+#include "Common/Web/HttpResponse.h"
 #include "Systems/Conversation/ConversationSystemHelpers.h"
-#include "Web/HttpResponse.h"
 
 namespace csp::multiplayer
 {
