+++ Library/src/Multiplayer/SignalR/SignalRClient.cpp
@@ -23,9 +23,9 @@
 #include "Multiplayer/WebSocketClient.h"
 
 #ifdef CSP_WASM
-#include "Web/EmscriptenWebClient/EmscriptenWebClient.h"
+#include "Common/Web/EmscriptenWebClient/EmscriptenWebClient.h"
 #else
-#include "Web/POCOWebClient/POCOWebClient.h"
+#include "Common/Web/POCOWebClient/POCOWebClient.h"
 #endif
 
 #include <stdexcept>
