+++ Library/src/Systems/Analytics/AnalyticsProviderGoogleUA.cpp
@@ -16,9 +16,9 @@
 #include "CSP/Systems/Analytics/AnalyticsProviderGoogleUA.h"
 
 #ifdef CSP_WASM
-#include "Web/EmscriptenWebClient/EmscriptenWebClient.h"
+#include "Common/Web/EmscriptenWebClient/EmscriptenWebClient.h"
 #else
-#include "Web/POCOWebClient/POCOWebClient.h"
+#include "Common/Web/POCOWebClient/POCOWebClient.h"
 #endif
 
 namespace
