+++ Library/src/Systems/GraphQL/GraphQL.cpp
@@ -15,8 +15,8 @@
  */
 #include "CSP/Systems/GraphQL/GraphQL.h"
 
+#include "Common/Web/GraphQLApi/GraphQLApi.h"
 #include "Services/ApiBase/ApiBase.h"
-#include "Web/GraphQLApi/GraphQLApi.h"
 
 using namespace csp::systems::graphqlservice;
