+++ Library/src/Systems/Spaces/SpaceSystem.cpp
@@ -19,6 +19,7 @@
 #include "CSP/CSPFoundation.h"
 #include "CSP/Common/SharedEnums.h"
 #include "CSP/Common/StringFormat.h"
+#include "CSP/Common/Web/HTTPResponseCodes.h"
 #include "CSP/Multiplayer/EventBus.h"
 #include "CSP/Multiplayer/MultiPlayerConnection.h"
 #include "CSP/Multiplayer/SpaceEntitySystem.h"
