+++ Library/src/Systems/Spatial/PointOfInterestInternalSystem.h
@@ -17,8 +17,8 @@
 
 #include "CSP/Systems/Spaces/Site.h"
 #include "CSP/Systems/Spatial/PointOfInterestSystem.h"
+#include "Common/Web/WebClient.h"
 #include "Debug/Logging.h"
-#include "Web/WebClient.h"
 
 namespace csp::systems
 {
