+++ Tests/src/PublicAPITests/SystemResultTests.cpp
@@ -28,13 +28,13 @@
 
 #include "gtest/gtest.h"
 #include <Common/CallHelpers.h>
-#include <Web/HttpRequest.h>
-#include <Web/WebClient.h>
+#include <Common/Web/HttpRequest.h>
+#include <Common/Web/WebClient.h>
 
 #ifdef CSP_WASM
-#include "Web/EmscriptenWebClient/EmscriptenWebClient.h"
+#include "Common/Web/EmscriptenWebClient/EmscriptenWebClient.h"
 #else
-#include "Web/POCOWebClient/POCOWebClient.h"
+#include "Common/Web/POCOWebClient/POCOWebClient.h"
 #endif
 
 using namespace csp::common;
