+++ Tests/src/PublicAPITests/UserSystemTests.cpp
@@ -21,10 +21,10 @@
 #include "CSP/Systems/SystemsManager.h"
 #include "CSP/Systems/Users/Profile.h"
 #include "CSP/Systems/Users/UserSystem.h"
+#include "Common/Web/HttpPayload.h"
 #include "SpaceSystemTestHelpers.h"
 #include "TestHelpers.h"
 #include "UserSystemTestHelpers.h"
-#include "Web/HttpPayload.h"
 
 #include "gtest/gtest.h"
 #include <filesystem>
